fix(textbox): test stories load to error
removed exported functions from storybook file to utils
